+++ svd-rs/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+- Add `dim_name` and `dim_array_index` to `DimElement`
 - Add `alternate_peripheral`, `prepend_to_name`, `append_to_name`,
   `header_struct_name` to `PeripheralInfo`, `alternate_cluster` to `ClusterInfo`
 - Add `protection` to `RegisterProperties` and `AddressBlock`